# backend/monolith/internal/auth/openapi/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package openapi

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	strictgin "github.com/oapi-codegen/runtime/strictmiddleware/gin"
)

const (
	CookieAuthScopes = "CookieAuth.Scopes"
)

// Defines values for PostAuthProviderCallbackParamsProvider.
const (
	Vk     PostAuthProviderCallbackParamsProvider = "vk"
	Yandex PostAuthProviderCallbackParamsProvider = "yandex"
)

// PostAuthProviderCallbackJSONBody defines parameters for PostAuthProviderCallback.
type PostAuthProviderCallbackJSONBody struct {
	// Code Authorization code received from the provider
	Code string `json:"code"`
}

// PostAuthProviderCallbackParamsProvider defines parameters for PostAuthProviderCallback.
type PostAuthProviderCallbackParamsProvider string

// PostProfileImageJSONBody defines parameters for PostProfileImage.
type PostProfileImageJSONBody struct {
	// Image Base64‑encoded image data (data URL format)
	Image string `json:"image"`
}

// PostAuthProviderCallbackJSONRequestBody defines body for PostAuthProviderCallback for application/json ContentType.
type PostAuthProviderCallbackJSONRequestBody PostAuthProviderCallbackJSONBody

// PostProfileImageJSONRequestBody defines body for PostProfileImage for application/json ContentType.
type PostProfileImageJSONRequestBody PostProfileImageJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Logout (alternative method)
	// (GET /auth/logout)
	GetAuthLogout(c *gin.Context)
	// Logout
	// (POST /auth/logout)
	PostAuthLogout(c *gin.Context)
	// OAuth2 callback endpoint
	// (POST /auth/{provider}/callback)
	PostAuthProviderCallback(c *gin.Context, provider PostAuthProviderCallbackParamsProvider)
	// Upload profile image
	// (POST /profile/image)
	PostProfileImage(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// GetAuthLogout operation middleware
func (siw *ServerInterfaceWrapper) GetAuthLogout(c *gin.Context) {
	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetAuthLogout(c)
}

// PostAuthLogout operation middleware
func (siw *ServerInterfaceWrapper) PostAuthLogout(c *gin.Context) {
	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAuthLogout(c)
}

// PostAuthProviderCallback operation middleware
func (siw *ServerInterfaceWrapper) PostAuthProviderCallback(c *gin.Context) {
	var err error

	// ------------- Path parameter "provider" -------------
	var provider PostAuthProviderCallbackParamsProvider

	err = runtime.BindStyledParameterWithOptions("simple", "provider", c.Param("provider"), &provider, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter provider: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostAuthProviderCallback(c, provider)
}

// PostProfileImage operation middleware
func (siw *ServerInterfaceWrapper) PostProfileImage(c *gin.Context) {
	c.Set(CookieAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostProfileImage(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/auth/logout", wrapper.GetAuthLogout)
	router.POST(options.BaseURL+"/auth/logout", wrapper.PostAuthLogout)
	router.POST(options.BaseURL+"/auth/:provider/callback", wrapper.PostAuthProviderCallback)
	router.POST(options.BaseURL+"/profile/image", wrapper.PostProfileImage)
}

type GetAuthLogoutRequestObject struct{}

type GetAuthLogoutResponseObject interface {
	VisitGetAuthLogoutResponse(w http.ResponseWriter) error
}

type GetAuthLogout200Response struct{}

func (response GetAuthLogout200Response) VisitGetAuthLogoutResponse(w http.ResponseWriter) error {
	w.WriteHeader(200)
	return nil
}

type GetAuthLogout401Response struct{}

func (response GetAuthLogout401Response) VisitGetAuthLogoutResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostAuthLogoutRequestObject struct{}

type PostAuthLogoutResponseObject interface {
	VisitPostAuthLogoutResponse(w http.ResponseWriter) error
}

type PostAuthLogout200JSONResponse struct {
	Status *string `json:"status,omitempty"`
}

func (response PostAuthLogout200JSONResponse) VisitPostAuthLogoutResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostAuthLogout401Response struct{}

func (response PostAuthLogout401Response) VisitPostAuthLogoutResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostAuthProviderCallbackRequestObject struct {
	Provider PostAuthProviderCallbackParamsProvider `json:"provider"`
	Body     *PostAuthProviderCallbackJSONRequestBody
}

type PostAuthProviderCallbackResponseObject interface {
	VisitPostAuthProviderCallbackResponse(w http.ResponseWriter) error
}

type PostAuthProviderCallback200JSONResponse struct {
	Status *string `json:"status,omitempty"`
	User   *struct {
		Email *string `json:"email,omitempty"`
		Id    *int64  `json:"id,omitempty"`
	} `json:"user,omitempty"`
}

func (response PostAuthProviderCallback200JSONResponse) VisitPostAuthProviderCallbackResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostAuthProviderCallback400Response struct{}

func (response PostAuthProviderCallback400Response) VisitPostAuthProviderCallbackResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostAuthProviderCallback401Response struct{}

func (response PostAuthProviderCallback401Response) VisitPostAuthProviderCallbackResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostAuthProviderCallback500Response struct{}

func (response PostAuthProviderCallback500Response) VisitPostAuthProviderCallbackResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

type PostProfileImageRequestObject struct {
	Body *PostProfileImageJSONRequestBody
}

type PostProfileImageResponseObject interface {
	VisitPostProfileImageResponse(w http.ResponseWriter) error
}

type PostProfileImage200JSONResponse struct {
	ImageUrl *string `json:"image_url,omitempty"`
}

func (response PostProfileImage200JSONResponse) VisitPostProfileImageResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostProfileImage400Response struct{}

func (response PostProfileImage400Response) VisitPostProfileImageResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type PostProfileImage401Response struct{}

func (response PostProfileImage401Response) VisitPostProfileImageResponse(w http.ResponseWriter) error {
	w.WriteHeader(401)
	return nil
}

type PostProfileImage500Response struct{}

func (response PostProfileImage500Response) VisitPostProfileImageResponse(w http.ResponseWriter) error {
	w.WriteHeader(500)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Logout (alternative method)
	// (GET /auth/logout)
	GetAuthLogout(ctx context.Context, request GetAuthLogoutRequestObject) (GetAuthLogoutResponseObject, error)
	// Logout
	// (POST /auth/logout)
	PostAuthLogout(ctx context.Context, request PostAuthLogoutRequestObject) (PostAuthLogoutResponseObject, error)
	// OAuth2 callback endpoint
	// (POST /auth/{provider}/callback)
	PostAuthProviderCallback(ctx context.Context, request PostAuthProviderCallbackRequestObject) (PostAuthProviderCallbackResponseObject, error)
	// Upload profile image
	// (POST /profile/image)
	PostProfileImage(ctx context.Context, request PostProfileImageRequestObject) (PostProfileImageResponseObject, error)
}

type (
	StrictHandlerFunc    = strictgin.StrictGinHandlerFunc
	StrictMiddlewareFunc = strictgin.StrictGinMiddlewareFunc
)

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetAuthLogout operation middleware
func (sh *strictHandler) GetAuthLogout(ctx *gin.Context) {
	var request GetAuthLogoutRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetAuthLogout(ctx, request.(GetAuthLogoutRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetAuthLogout")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(GetAuthLogoutResponseObject); ok {
		if err := validResponse.VisitGetAuthLogoutResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostAuthLogout operation middleware
func (sh *strictHandler) PostAuthLogout(ctx *gin.Context) {
	var request PostAuthLogoutRequestObject

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostAuthLogout(ctx, request.(PostAuthLogoutRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostAuthLogout")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostAuthLogoutResponseObject); ok {
		if err := validResponse.VisitPostAuthLogoutResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostAuthProviderCallback operation middleware
func (sh *strictHandler) PostAuthProviderCallback(ctx *gin.Context, provider PostAuthProviderCallbackParamsProvider) {
	var request PostAuthProviderCallbackRequestObject

	request.Provider = provider

	var body PostAuthProviderCallbackJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostAuthProviderCallback(ctx, request.(PostAuthProviderCallbackRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostAuthProviderCallback")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostAuthProviderCallbackResponseObject); ok {
		if err := validResponse.VisitPostAuthProviderCallbackResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}

// PostProfileImage operation middleware
func (sh *strictHandler) PostProfileImage(ctx *gin.Context) {
	var request PostProfileImageRequestObject

	var body PostProfileImageJSONRequestBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.Status(http.StatusBadRequest)
		ctx.Error(err)
		return
	}
	request.Body = &body

	handler := func(ctx *gin.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostProfileImage(ctx, request.(PostProfileImageRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostProfileImage")
	}

	response, err := handler(ctx, request)

	if err != nil {
		ctx.Error(err)
		ctx.Status(http.StatusInternalServerError)
	} else if validResponse, ok := response.(PostProfileImageResponseObject); ok {
		if err := validResponse.VisitPostProfileImageResponse(ctx.Writer); err != nil {
			ctx.Error(err)
		}
	} else if response != nil {
		ctx.Error(fmt.Errorf("unexpected response type: %T", response))
	}
}